Added multiple keywords that link to the same command.
Difficulty: Tried to change the COMMAND_WORD to a string array and add all the relevant key words but broke the parser.
Solution: Made an alternative COMMAND_WORDS for the string array and redid the parser to search the array for the command word.

Difficulty: Help message was not updated with the various keywords.
Solution 1 (failed) : Use COMMAND_WORD = concatenate(COMMAND_WORDS) with "-OR-" in between, ended up breaking many tests
Solution 2: revert COMMAND_WORD to the original word and use the concatenate function explicitly in each command message