Refactor: Move color from UI component to Model Component, fixed issue of not auto refreshing
Each unique tag has a color attached.

I did this by exposing the UI to model, so that when then the tags' string color change, I can call the personPanelList to reset and refresh.

In the future if you need to excess personPanelList in model, you can call resetPanel in UiManager.

Updated UserGuide to reflect changes for the user.

TODO:

Allow users to set a color to a tag