Extreme overhaul for appointments.
Now appointments are is a list under a person, which means that each person can have multiple appointments.

Appointments now have descriptions.

Removed off appointments for now as it is difficult for users to specify which apppointment to off.

Removed editing of appointment in edit because it doesnt make sense.

Todo:
A panel for just appointments so that users can easily delete/edit appointments